[ItemOption 도메인 추가]

+++ src/main/java/com/example/cih/dto/cart/CartDTO.java
@@ -15,5 +15,5 @@ public class CartDTO {
     private Long shopItemId;
     private String itemName;
     private int itemCount;
-    private int itemOption;
+    private Long itemOption;
 }
